Got more tests passing. There's only one failure now!
There was a silly problem where we weren't passing arguments into getMany.

There was a really subtle problem in the stub graph: the internal toId property
on the relationship object was conflicting with the toId property from opinionData.

We were having a hard time diagnosing these issues because the service-layer tests
were not running, due to earlier failures in the integration tests. Excluding the
integration tests revealed failures in the service layer tests.