Delete integration tests. Rename api/ to routes/.
We verified that all the functionality covered in the api tests are
now covered in the unit tests, so we are removing them. We felt like
the only major missing piece was to test the routing, so we previously
added those tests, and now we are renaming the directory they were
contained in to user a more descriptive name.